[AIP-21] Move contrib.sensors.file_sensor sensors.filesystem
