+++ airflow/contrib/sensors/file_sensor.py
@@ -16,54 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
+"""This module is deprecated. Please use `airflow.sensors.filesystem`."""
 
-import os
-from glob import glob
+import warnings
 
-from airflow.hooks.filesystem import FSHook
-from airflow.sensors.base_sensor_operator import BaseSensorOperator
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.sensors.filesystem import FileSensor  # noqa
 
-
-class FileSensor(BaseSensorOperator):
-    """
-    Waits for a file or folder to land in a filesystem.
-
-    If the path given is a directory then this sensor will only return true if
-    any files exist inside it (either directly, or within a subdirectory)
-
-    :param fs_conn_id: reference to the File (path)
-        connection id
-    :type fs_conn_id: str
-    :param filepath: File or folder name (relative to
-        the base path set within the connection), can be a glob.
-    :type fs_conn_id: str
-    """
-    template_fields = ('filepath',)
-    ui_color = '#91818a'
-
-    @apply_defaults
-    def __init__(self,
-                 filepath,
-                 fs_conn_id='fs_default',
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.filepath = filepath
-        self.fs_conn_id = fs_conn_id
-
-    def poke(self, context):
-        hook = FSHook(self.fs_conn_id)
-        basepath = hook.get_path()
-        full_path = os.path.join(basepath, self.filepath)
-        self.log.info('Poking for file %s', full_path)
-
-        for path in glob(full_path):
-            if os.path.isfile(path):
-                return True
-
-            for _, _, files in os.walk(full_path):
-                if len(files) > 0:
-                    return True
-        return False
+warnings.warn(
+    "This module is deprecated. Please use `airflow.sensors.filesystem`.",
+    DeprecationWarning, stacklevel=2
+)
